moving addUserDataToChats function to utils for reusability
